requirements.txt
pymongo~=4.5.0
bcrypt~=4.0.1
cryptography~=41.0.3
scapy~=2.5.0
python-pytun~=2.4.1
pycryptodome~=3.19.0